Revert "Vision api"
