+++ src/Containers/homescreen/PredictionModal.js
@@ -7,6 +7,11 @@ const PredictionModal = props => {
 
   const navigation = useNavigation();
 
+  const toArray = (predictions) => {
+    const predictionArray = predictions.split(",");
+    return predictionArray;
+  }
+
   return (
     <Modal
       animationType="slide"
@@ -19,7 +24,7 @@ const PredictionModal = props => {
           <Text style={styles.modalText}>Choose a name:</Text>
           <List>
             {
-              props.predictions.map(prediction => {
+              toArray(props.predictions).map(prediction => {
                 return (
                   <ListItem
                     style={styles.item}
