+++ src/redux/reducer.js
@@ -1,5 +1,5 @@
 import { combineReducers } from "redux";
-import { ADD_CATEGORY, REMOVE_CATEGORY, ORDER_CATEGORY } from "./actions";
+import { ADD_CATEGORY, REMOVE_CATEGORY } from "./actions";
 
 const categoryReducer = (state = ["Food"], action) => {
   switch (action.type) {
@@ -7,8 +7,6 @@ const categoryReducer = (state = ["Food"], action) => {
       return [...state, action.payload];
     case REMOVE_CATEGORY:
       return state.filter((category) => category !== action.payload);
-    case ORDER_CATEGORY:
-      return action.payload;
     default:
       return state;
   }
