+++ yarn-error.log
@@ -1,8 +1,8 @@
 Arguments: 
-  C:\Program Files\nodejs\node.exe C:\Program Files (x86)\Yarn\bin\yarn.js add react-native-floating action react-native-draggable-flatlist react-native-swipeable-item react-native-reanimated
+  C:\Program Files\nodejs\node.exe C:\Program Files (x86)\Yarn\bin\yarn.js add @react-native-community/async-storage @tensorflow/tfjs @tensorflow/tfjs-react-native expo-gl @tensorflow-models/mobilenet jpeg-js
 
 PATH: 
-  C:\Python38\Scripts\;C:\Python38\;C:\Program Files\Java\jdk-11.0.6\bin;C:\Program Files (x86)\Vim\vim82;c:\WINDOWS\System32\WBem;C:\ProgramData\chocolatey\bin;C:\Windows\System32\;C:\Program Files\nodejs\;C:\Program Files (x86)\Yarn\bin\;C:\Users\nicho\AppData\Roaming\Python\Python38\Scripts;C:\Program Files\Git\cmd;C:\Users\nicho\AppData\Local\Microsoft\WindowsApps;C:\Users\nicho\AppData\Local\Programs\Microsoft VS Code\bin;C:\Users\nicho\AppData\Local\GitHubDesktop\bin;C:\Users\nicho\AppData\Roaming\npm;C:\Users\nicho\AppData\Local\Yarn\bin
+  C:\Python38\Scripts\;C:\Python38\;C:\Program Files\Java\jdk-11.0.6\bin;C:\Program Files (x86)\Vim\vim82;c:\WINDOWS\System32\WBem;C:\ProgramData\chocolatey\bin;C:\Windows\System32\;C:\Program Files\Git\cmd;C:\Program Files\nodejs\;C:\Program Files (x86)\Yarn\bin\;C:\Users\nicho\AppData\Local\Microsoft\WindowsApps;C:\Users\nicho\AppData\Local\Programs\Microsoft VS Code\bin;C:\Users\nicho\AppData\Local\GitHubDesktop\bin;C:\Users\nicho\AppData\Roaming\npm;C:\Users\nicho\AppData\Local\Yarn\bin
 
 Yarn version: 
   1.22.4
@@ -14,17 +14,7 @@ Platform:
   win32 x64
 
 Trace: 
-  Error: https://registry.yarnpkg.com/react-native-floating: Not found
-      at Request.params.callback [as _callback] (C:\Program Files (x86)\Yarn\lib\cli.js:66096:18)
-      at Request.self.callback (C:\Program Files (x86)\Yarn\lib\cli.js:140748:22)
-      at Request.emit (events.js:315:20)
-      at Request.<anonymous> (C:\Program Files (x86)\Yarn\lib\cli.js:141720:10)
-      at Request.emit (events.js:315:20)
-      at IncomingMessage.<anonymous> (C:\Program Files (x86)\Yarn\lib\cli.js:141642:12)
-      at Object.onceWrapper (events.js:421:28)
-      at IncomingMessage.emit (events.js:327:22)
-      at endReadableNT (_stream_readable.js:1225:12)
-      at processTicksAndRejections (internal/process/task_queues.js:84:21)
+  Error: EPERM: operation not permitted, unlink 'C:\Users\nicho\Documents\Censible\node_modules\@babel\compat-data\node_modules\.bin'
 
 npm manifest: 
   {
@@ -48,9 +38,8 @@ npm manifest:
       "@tensorflow/tfjs-react-native": "^0.3.0",
       "core-js": "^3.6.5",
       "expo": "^37.0.11",
-      "expo-camera": "~8.2.0",
+      "expo-camera": "^8.2.0",
       "expo-constants": "~9.0.0",
-      "expo-file-system": "~8.1.0",
       "expo-font": "~8.1.0",
       "expo-gl": "^8.2.0",
       "expo-image-picker": "~8.1.0",
@@ -76,10 +65,6 @@ npm manifest:
       "react-native-screens": "^2.2.0",
       "react-native-svg": "^12.1.0",
       "react-native-web": "~0.11.7",
-      "react-redux": "^7.2.0",
-      "redux": "^4.0.5",
-      "redux-logger": "^3.0.6",
-      "redux-persist": "^6.0.0",
       "websocket-extensions": "^0.1.4"
     },
     "devDependencies": {
@@ -2907,11 +2892,6 @@ Lockfile:
     dependencies:
       is-obj "^1.0.0"
   
-  deep-diff@^0.3.5:
-    version "0.3.8"
-    resolved "https://registry.yarnpkg.com/deep-diff/-/deep-diff-0.3.8.tgz#c01de63efb0eec9798801d40c7e0dae25b582c84"
-    integrity sha1-wB3mPvsO7JeYgB1Ax+Da4ltYLIQ=
-  
   deep-extend@^0.6.0:
     version "0.6.0"
     resolved "https://registry.yarnpkg.com/deep-extend/-/deep-extend-0.6.0.tgz#c4fa7c95404a17a9c3e8ca7e1537312b736330ac"
@@ -3203,7 +3183,7 @@ Lockfile:
       path-browserify "^1.0.0"
       url-parse "^1.4.4"
   
-  expo-camera@~8.2.0:
+  expo-camera@^8.2.0:
     version "8.2.0"
     resolved "https://registry.yarnpkg.com/expo-camera/-/expo-camera-8.2.0.tgz#9c849f0dc409e1fe6ee886f479e0ee0050b6d386"
     integrity sha512-wGVqQVVR4YCkHuI7u+qloI1p2UbYwpU8hs0IEM+1Th4JQekAPUEfOmJI6ZHGj36El8f87ghRed2MXzOssn4hHg==
@@ -3819,7 +3799,7 @@ Lockfile:
     resolved "https://registry.yarnpkg.com/hoist-non-react-statics/-/hoist-non-react-statics-2.5.5.tgz#c5903cf409c0dfd908f388e619d86b9c1174cb47"
     integrity sha512-rqcy4pJo55FTTLWt+bU8ukscqHeE/e9KWvsOW2b/a3afxQZhwkQdT1rPPCJ0rYXdj4vNcasY8zHTH+jF/qStxw==
   
-  hoist-non-react-statics@^3.1.0, hoist-non-react-statics@^3.3.0, hoist-non-react-statics@^3.3.2:
+  hoist-non-react-statics@^3.1.0, hoist-non-react-statics@^3.3.2:
     version "3.3.2"
     resolved "https://registry.yarnpkg.com/hoist-non-react-statics/-/hoist-non-react-statics-3.3.2.tgz#ece0acaf71d62c2969c2ec59feff42a4b1a85b45"
     integrity sha512-/gGivxi8JPKWNm/W0jSmzcMPpfpPLc3dY/6GxhX2hQ9iGj3aDfklV4ET7NjKpSinLpJ5vafa9iiGIEZg10SfBw==
@@ -5762,7 +5742,7 @@ Lockfile:
       prop-types "^15.6.2"
       scheduler "^0.15.0"
   
-  react-is@^16.12.0, react-is@^16.13.0, react-is@^16.7.0, react-is@^16.8.0, react-is@^16.8.1, react-is@^16.8.4, react-is@^16.9.0:
+  react-is@^16.12.0, react-is@^16.13.0, react-is@^16.7.0, react-is@^16.8.0, react-is@^16.8.1, react-is@^16.8.4:
     version "16.13.1"
     resolved "https://registry.yarnpkg.com/react-is/-/react-is-16.13.1.tgz#789729a4dc36de2999dc156dd6c1d9c18cea56a4"
     integrity sha512-24e6ynE2H+OKt4kqsOvNd8kBpV65zoxbA4BVsEOB3ARVWQki/DHzaUoC5KuON/BiccDaCCTZBuOcfZs70kR8bQ==
@@ -5973,17 +5953,6 @@ Lockfile:
       stacktrace-parser "^0.1.3"
       whatwg-fetch "^3.0.0"
   
-  react-redux@^7.2.0:
-    version "7.2.0"
-    resolved "https://registry.yarnpkg.com/react-redux/-/react-redux-7.2.0.tgz#f970f62192b3981642fec46fd0db18a074fe879d"
-    integrity sha512-EvCAZYGfOLqwV7gh849xy9/pt55rJXPwmYvI4lilPM5rUT/1NxuuN59ipdBksRVSvz0KInbPnp4IfoXJXCqiDA==
-    dependencies:
-      "@babel/runtime" "^7.5.5"
-      hoist-non-react-statics "^3.3.0"
-      loose-envify "^1.4.0"
-      prop-types "^15.7.2"
-      react-is "^16.9.0"
-  
   react-refresh@^0.4.0:
     version "0.4.2"
     resolved "https://registry.yarnpkg.com/react-refresh/-/react-refresh-0.4.2.tgz#54a277a6caaac2803d88f1d6f13c1dcfbd81e334"
@@ -6050,26 +6019,6 @@ Lockfile:
       safe-buffer "~5.1.1"
       string_decoder "~1.1.1"
       util-deprecate "~1.0.1"
-  
-  redux-logger@^3.0.6:
-    version "3.0.6"
-    resolved "https://registry.yarnpkg.com/redux-logger/-/redux-logger-3.0.6.tgz#f7555966f3098f3c88604c449cf0baf5778274bf"
-    integrity sha1-91VZZvMJjzyIYExEnPC69XeCdL8=
-    dependencies:
-      deep-diff "^0.3.5"
-  
-  redux-persist@^6.0.0:
-    version "6.0.0"
-    resolved "https://registry.yarnpkg.com/redux-persist/-/redux-persist-6.0.0.tgz#b4d2972f9859597c130d40d4b146fecdab51b3a8"
-    integrity sha512-71LLMbUq2r02ng2We9S215LtPu3fY0KgaGE0k8WRgl6RkqxtGfl7HUozz1Dftwsb0D/5mZ8dwAaPbtnzfvbEwQ==
-  
-  redux@^4.0.5:
-    version "4.0.5"
-    resolved "https://registry.yarnpkg.com/redux/-/redux-4.0.5.tgz#4db5de5816e17891de8a80c424232d06f051d93f"
-    integrity sha512-VSz1uMAH24DM6MF72vcojpYPtrTUu3ByVWfPL1nPfVRb5mZVTve5GnNCUV53QM/BZ66xfWrm0CTWoM+Xlz8V1w==
-    dependencies:
-      loose-envify "^1.4.0"
-      symbol-observable "^1.2.0"
   
   regenerate-unicode-properties@^8.2.0:
     version "8.2.0"
@@ -6703,11 +6652,6 @@ Lockfile:
     version "1.0.1"
     resolved "https://registry.yarnpkg.com/symbol-observable/-/symbol-observable-1.0.1.tgz#8340fc4702c3122df5d22288f88283f513d3fdd4"
     integrity sha1-g0D8RwLDEi310iKI+IKD9RPT/dQ=
-  
-  symbol-observable@^1.2.0:
-    version "1.2.0"
-    resolved "https://registry.yarnpkg.com/symbol-observable/-/symbol-observable-1.2.0.tgz#c22688aed4eab3cdc2dfeacbb561660560a00804"
-    integrity sha512-e900nM8RRtGhlV36KGEU9k65K3mPb1WV70OdjfxlG2EAuM1noi/E/BaW/uMhL7bPEssK8QV57vN3esixjUvcXQ==
   
   tar@^4.4.2:
     version "4.4.13"
